# server/openapi.yaml
openapi: 3.0.3
info:
  title: Risk Register API
  version: 1.0.0
  description: >-
    REST-API zur Verwaltung eines Risikoregisters.


    Die API unterstützt das Erfassen, Suchen, Aktualisieren und Löschen von
    Risiken.

    Risikowerte liegen zwischen 1 und 25 und ergeben sich fachlich aus
    Eintrittswahrscheinlichkeit × Auswirkung.
  contact:
    name: Risk Register Team
  license:
    name: Private use
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: signed-token
  schemas:
...
    RiskPage:
      title: RiskPage
      description: Paginierte Ergebnisliste für Risiken.
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Risk"
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 500
          example: 25
        total:
          type: integer
          minimum: 0
          example: 42
        totalPages:
          type: integer
          minimum: 1
          example: 2
      required:
        - items
        - page
        - pageSize
        - total
        - totalPages
    ErrorResponse:
      title: ErrorResponse
      description: Einheitliche Fehlerantwort der Risk-Register-API.
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - RISK_NOT_FOUND
            - MEASURE_NOT_FOUND
            - USER_NOT_FOUND
            - AUTH_FAILED
            - FORBIDDEN
            - INTERNAL_ERROR
          example: RISK_NOT_FOUND
        message:
          type: string
          example: Das Risiko mit der angegebenen ID wurde nicht gefunden.
      required:
        - code
        - message
...
paths:
...
  /api/risks:
    get:
      operationId: listRisks
      summary: Risiken abrufen
      tags:
        - Risiken
      description: Liefert Risiken serverseitig gefiltert, sortiert und paginiert.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
          description: Eins-basierte Seitennummer.
        - schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          in: query
          name: pageSize
          required: false
          description: Anzahl der Datensätze pro Seite.
        - schema:
            type: string
            maxLength: 200
          example: Lieferkette
          in: query
          name: search
          required: false
          description: Freitextsuche in Referenz, Bezeichnung, Beschreibung, Kategorie und
            Verantwortlichem.
        - schema:
            type: string
            enum:
              - Offen
              - In Bearbeitung
              - Überwacht
              - Geschlossen
          in: query
          name: status
          required: false
          description: Filtert nach Bearbeitungsstatus.
        - schema:
            type: string
            maxLength: 200
          in: query
          name: category
          required: false
          description: Filtert nach exakt passender Kategorie.
        - schema:
            type: string
            maxLength: 200
          in: query
          name: owner
          required: false
          description: Filtert nach exakt passender verantwortlicher Person.
        - schema:
            type: string
            enum:
              - reference
              - title
              - category
              - owner
              - currentScore
              - status
              - dueDate
              - reviewDate
              - createdAt
              - updatedAt
            default: currentScore
          in: query
          name: sortBy
          required: false
          description: Sortierfeld.
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          in: query
          name: sortDirection
          required: false
          description: Sortierrichtung.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Die gefundene Risikoseite.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskPage"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createRisk
      summary: Ein Risiko erfassen
      tags:
        - Risiken
      description: Erfasst ein neues Risiko. ID, Referenz und Zeitstempel werden
        serverseitig erzeugt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRisk"
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Das erfolgreich erfasste Risiko.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/risks/{id}:
    get:
      operationId: getRisk
      summary: Ein Risiko abrufen
      tags:
        - Risiken
      description: Liefert ein einzelnes Risiko anhand seiner technischen ID.
      parameters:
        - schema:
            type: string
            minLength: 1
          example: 01J00000000000000000000001
          in: path
          name: id
          required: true
          description: Technische ID des Risikos.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Das angeforderte Risiko.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Risiko gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRisk
      summary: Ein Risiko aktualisieren
      tags:
        - Risiken
      description: Aktualisiert die übergebenen Felder eines vorhandenen Risikos.
        Nicht übergebene Felder bleiben unverändert.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRisk"
      parameters:
        - schema:
            type: string
            minLength: 1
          example: 01J00000000000000000000001
          in: path
          name: id
          required: true
          description: Technische ID des Risikos.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Das aktualisierte Risiko.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Risiko gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRisk
      summary: Ein Risiko zur Löschung markieren
      tags:
        - Risiken
      description: Markiert ein Risiko als gelöscht. Es bleibt in SQLite erhalten,
        wird aber in aktiven Listen nicht mehr ausgeliefert.
      parameters:
        - schema:
            type: string
            minLength: 1
          example: 01J00000000000000000000001
          in: path
          name: id
          required: true
          description: Technische ID des Risikos.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Das Risiko wurde erfolgreich zur Löschung markiert.
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Risiko gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/measures:
    get:
      operationId: listMeasures
      summary: Maßnahmen abrufen
      tags:
        - Maßnahmen
      description: Liefert Maßnahmen serverseitig gefiltert, sortiert und paginiert.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
          description: Eins-basierte Seitennummer.
        - schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          in: query
          name: pageSize
          required: false
          description: Anzahl der Datensätze pro Seite.
        - schema:
            type: string
            maxLength: 200
          in: query
          name: search
          required: false
          description: Freitextsuche in Titel, Beschreibung und Verantwortlichem.
        - schema:
            type: string
            enum:
              - Offen
              - In Arbeit
              - Erledigt
          in: query
          name: status
          required: false
          description: Filtert nach Bearbeitungsstatus.
        - schema:
            type: string
            enum:
              - Kritisch
              - Hoch
              - Normal
          in: query
          name: priority
          required: false
          description: Filtert nach Priorität.
        - schema:
            type: string
            maxLength: 200
          in: query
          name: owner
          required: false
          description: Filtert nach exakt passender verantwortlicher Person.
        - schema:
            type: string
            minLength: 1
          in: query
          name: riskId
          required: false
          description: Filtert nach verknüpftem Risiko.
        - schema:
            type: string
            enum:
              - title
              - owner
              - dueDate
              - priority
              - status
              - createdAt
              - updatedAt
            default: dueDate
          in: query
          name: sortBy
          required: false
          description: Sortierfeld.
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          in: query
          name: sortDirection
          required: false
          description: Sortierrichtung.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Die gefundene Maßnahmenseite.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurePage"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMeasure
      summary: Eine Maßnahme anlegen
      tags:
        - Maßnahmen
      description: Legt eine neue Maßnahme optional mit Bezug zu einem Risiko an.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeasure"
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Die erfolgreich angelegte Maßnahme.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measure"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Maßnahmendaten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/measures/{id}:
    get:
      operationId: getMeasure
      summary: Eine Maßnahme abrufen
      tags:
        - Maßnahmen
      description: Liefert eine einzelne Maßnahme anhand ihrer technischen ID.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
          description: Technische ID der Maßnahme.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Die angeforderte Maßnahme.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measure"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde keine Maßnahme gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMeasure
      summary: Eine Maßnahme bearbeiten
      tags:
        - Maßnahmen
      description: Aktualisiert die übergebenen Felder einer vorhandenen Maßnahme.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeasure"
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
          description: Technische ID der Maßnahme.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Die aktualisierte Maßnahme.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measure"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Maßnahmendaten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde keine Maßnahme gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeasure
      summary: Eine Maßnahme löschen
      tags:
        - Maßnahmen
      description: Entfernt eine Maßnahme nach Benutzerbestätigung in der Oberfläche.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
          description: Technische ID der Maßnahme.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Die Maßnahme wurde gelöscht.
        "401":
          description: Die Anfrage benötigt eine gültige Anmeldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde keine Maßnahme gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/profile:
    get:
      operationId: getProfile
      summary: Eigenes Profil abrufen
      tags:
        - Benutzer
      description: Liefert das Profil des angemeldeten Benutzers ohne Passwortdaten.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Benutzer gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      summary: Eigenes Profil bearbeiten
      tags:
        - Benutzer
      description: Aktualisiert Profilfelder oder Passwort des angemeldeten Benutzers.
        Rolle und Aktivstatus bleiben Admins vorbehalten.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfile"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Benutzer gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users:
    get:
      operationId: listUsers
      summary: Benutzer auflisten
      tags:
        - Benutzer
      description: Liefert alle Benutzer ohne Passwortdaten.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Nur Admins dürfen Benutzer verwalten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      summary: Benutzer anlegen
      tags:
        - Benutzer
      description: Legt einen Benutzer mit gehashtem Passwort an.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Nur Admins dürfen Benutzer verwalten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
servers:
  - url: http://127.0.0.1:8000
    description: Lokale Entwicklungsumgebung
...